Changement image, corrections bugs, début préparation de commandes

+++ fr.univcotedazur.polytech.si4.fsm.2020.project/src-gen/fr/univcotedazur/polytech/si4/fsm/project/fsm/FSMStateMachineListener.java
@@ -23,12 +23,12 @@ public class FSMStateMachineListener implements SCInterfaceListener, SCInterface
 	}
 
 	@Override
-	public double newPrice() {
+	public long newPrice() {
 		return machineController.newPrice();
 	}
 
 	@Override
-	public double currentMoney() {
+	public long currentMoney() {
 		return machineController.currentMoney();
 	}
 }
